moved the additional function back to data.py - need to rename this script to manage_hashes
